requirements.txt
# Core ML Frameworks
unsloth
transformers==4.56.2
torch>=2.0.0

# Training and Data
datasets>=2.14.0
trl==0.22.2

# HuggingFace and Model Management
huggingface_hub>=0.16.0
accelerate>=0.20.0

# Image Processing
Pillow>=9.0.0

# Text Processing and Tokenization
sentencepiece>=0.1.99
protobuf>=3.20.0

# Utilities
einops>=0.6.0
addict>=2.4.0
easydict>=1.10
PyYAML>=6.0

# Evaluation Metrics
jiwer>=3.0.0

# Optional: For better performance
# bitsandbytes>=0.41.0
# flash-attn>=2.0.0
